Fix import lint warning and shift bases tests
