+++ tests/eko/io/test_runcards.py
@@ -1,5 +1,4 @@
 import copy
-from dataclasses import fields
 from io import StringIO
 
 import numpy as np
@@ -11,42 +10,6 @@ from eko import interpolation
 from eko.io import runcards as rc
 from eko.io.bases import Bases
 from ekomark.data.operators import default_card as operator_card
-
-
-class TestBases:
-    XGRID_TEST = [1e-3, 1e-2, 1e-1, 1.0]
-
-    def test_serialization(self):
-        rot = Bases(interpolation.XGrid(self.XGRID_TEST))
-
-        d = rot.raw
-        rot1 = rot.from_dict(d)
-
-        for f in fields(Bases):
-            assert getattr(rot, f.name) == getattr(rot1, f.name)
-
-        assert d["targetgrid"] is None
-        assert "_targetgrid" not in d
-
-    def test_pids(self):
-        rot = Bases(interpolation.XGrid(self.XGRID_TEST))
-
-        # no check on correctness of value set
-        rot.inputpids = [0, 1]
-        # but the internal grid is unmodified
-        assert len(rot.pids) == 14
-        # and fallback implemented for unset external bases
-        assert np.all(rot.targetpids == rot.pids)
-
-    def test_grids(self):
-        rot = Bases(interpolation.XGrid(self.XGRID_TEST))
-
-        # no check on correctness of value set
-        rot.inputgrid = interpolation.XGrid([0.1, 1])
-        # but the internal grid is unmodified
-        assert len(rot.xgrid) == len(self.XGRID_TEST)
-        # and fallback implemented for unset external grids
-        assert np.all(rot.targetgrid == rot.xgrid)
 
 
 def test_flavored_mu2grid():
